Split Github and LDAP tests into separate files
